Fix alert discover order convert PollWirelessChannelAsFrequency to a trait and fix it up
