+++ LibreNMS/OS/Arubaos.php
@@ -42,6 +42,8 @@ class Arubaos extends OS implements
     WirelessPowerDiscovery,
     WirelessUtilizationDiscovery
 {
+    use OS\Traits\PollWirelessChannelAsFrequency;
+
     /**
      * Discover wireless client counts. Type is clients.
      * Returns an array of LibreNMS\Device\Sensor objects that have been discovered
